Create Disable_cmd.ps1
# README

This PowerShell script is designed to disable the Command Prompt through the registry to prevent unauthorized access and protect against bad USB attacks that send commands using the command line. The script requires a password to be entered before the Command Prompt can be used, which helps to ensure that only authorized users can access the Command Prompt.

## Getting Started

To use this script, follow these steps:

1. Open PowerShell as an administrator.

2. Copy and paste the script into the PowerShell console.

3. Edit the password variable in the script to set a secure password that will be required to use the Command Prompt. Make sure to use a strong password and keep it secret.

4. Run the script.

## Usage

When the script is run, it will disable the Command Prompt through the registry and prompt the user for a password. To use the Command Prompt, the user must enter the correct password when prompted. Once the password is verified, the Command Prompt will be enabled and the user can use it as normal. When the user is finished using the Command Prompt, they can close it to disable it again.

## Password Security

The password is stored in an encrypted format in a file called "securepassword.xml" in the root of the C: drive. The file is protected by the Windows file system, but it's important to ensure that appropriate access controls are in place to prevent unauthorized access to the file. To change the password, edit the script to set a new password and run the script again.

## Disclaimer

This script is provided as-is and is intended for educational purposes only. Use at your own risk. The author is not responsible for any damages or losses that may occur as a result of using this script. Always use good security practices to protect your systems and data.